Experiment/requirements.txt
numpy==1.22.3
Pillow==9.1.0
scikit_image==0.19.2
skimage==0.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
